unsigned long -> uint32

+++ common/canvas.h
@@ -2,16 +2,19 @@
 #define __CANVAS_H__
 
 #include "rect.h"
+#include "types.h"
+
 
 class Icon;
+
 
 class Canvas
 {
 public:
     virtual void erase(const Rect &rect) = 0;
-    virtual void drawFrame(const Rect &rect, unsigned long color) = 0;
+    virtual void drawFrame(const Rect &rect, uint32 color) = 0;
     virtual void drawText(const char *text, const Rect &rect,
-                          unsigned long fg, unsigned long bg) = 0;
+                          uint32 fg, uint32 bg) = 0;
     virtual void drawIcon(Icon *icon, int x, int y) = 0;
 };
