+++ x11/x11_canvas.h
@@ -14,13 +14,13 @@ public:
     ~X11Canvas();
 
     virtual void erase(const Rect &rect);
-    virtual void drawFrame(const Rect &rect, unsigned long color);
+    virtual void drawFrame(const Rect &rect, uint32 color);
     virtual void drawText(const char *text, const Rect &rect,
-                          unsigned long fg, unsigned long bg);
+                          uint32 fg, uint32 bg);
     virtual void drawIcon(Icon *icon, int x, int y);
 
-    void fillRectangle(const Rect &rect, unsigned long color);
-    void drawLine(int x1, int y1, int x2, int y2, unsigned long color);
+    void fillRectangle(const Rect &rect, uint32 color);
+    void drawLine(int x1, int y1, int x2, int y2, uint32 color);
 
     GC gc() { return _gc; }
