+++ x11/x11_icon.h
@@ -2,6 +2,7 @@
 #define __X11_ICON_H__
 
 #include "icon.h"
+#include "types.h"
 
 #include <X11/Xlib.h>
 
@@ -13,7 +14,7 @@ class X11Icon : public Icon
 {
 public:
     X11Icon(int width, int height, X11Widget *parent);
-    X11Icon(int width, int height, X11Widget *parent, const unsigned long *argb_data, unsigned long bg_color);
+    X11Icon(int width, int height, X11Widget *parent, const uint32 *argb_data, uint32 bg_color);
     X11Icon(const X11Icon &other);
     ~X11Icon();
     X11Icon &operator=(const X11Icon &other);
